Renamed parser-types.ts
